% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcPetToAllHrus}
\alias{calcPetToAllHrus}
\title{Calculates the values of Potential evapotranspiration on all singleHrus taking constant latitude}
\usage{
calcPetToAllHrus(dHRUM_ptr, Latitude, PetTypeStr)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{Latitude}{single number for Oudin method}

\item{PetTypeStr}{variable on selection of PET models}
}
\description{
Setting of Pet method is done by \code{PetTypeStr}, methods implemented are:
\code{OUDIN}, \code{HAMON}, \code{THORNTHWAITE}, \code{BLANEYCRIDDLE}, \code{JENSENHAISE}, \code{MCGUINNESSBORDNE}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTDateInputsToAlldHrus(dhrus, Prec = prec, Temp = temp,
  DateVec = as.Date(c("1990/01/30","1990/01/31","1990/02/01")))
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
setParamsToAlldHrus(dHRUM_ptr = dhrus,as.numeric(ParDF[1,]),names(ParDF))
calcPetToAllHrus(dHRUM_ptr = dhrus,50.1,"HAMON")
calcHBInAlldHrus(dHRUM_ptr = dhrus)
}
