% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcPetToAllHrusDist}
\alias{calcPetToAllHrusDist}
\title{Calculates the values of Potential evapotranspiration on all singleHrus}
\usage{
calcPetToAllHrusDist(dHRUM_ptr, Latitude, PetTypeStrNames, HruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{Latitude}{vector for latitude based methods}

\item{PetTypeStrNames}{char vector on selection of PET models}

\item{HruIds}{vector of Hru Ids}
}
\description{
Setting of Pet method is done by \code{PetTypeStr}, methods implemented are:
\code{OUDIN}, \code{HAMON}, \code{THORNTHWAITE}, \code{BLANEYCRIDDLE}, \code{JENSENHAISE}, \code{MCGUINNESSBORDNE}
}
\examples{
nHrus <- 20
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTDateInputsToAlldHrus(dhrus, Prec = prec, Temp = temp,
  DateVec = as.Date(c("1990/01/30","1990/01/31","1990/02/01")))
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 1,  KS = 0.01, KF = 0.03, ADIV = 0.8, CDIV = 0.3,
                      SDIV = 0.3, CAN_ST = 1., STEM_ST = 1., CSDIV = 0.8, TETR = 0, DDFA = 0.75, TMEL = 0.0,
                      RETCAP = 10 ,L = 0.5, D_BYPASS = 0.8, B_EXP=1, KS2 = 0.1, THR = 10, ALPHA =0.5)
setParamsToAlldHrus(dHRUM_ptr = dhrus,as.numeric(ParDF[1,]),names(ParDF))
latits = runif(nHrus)
pets =c("OUDIN", "HAMON", "THORNTHWAITE","BLANEYCRIDDLE","JENSENHAISE", "MCGUINNESSBORDNE")
PetsNams = rep(pets,times=4)
calcPetToAllHrusDist(dHRUM_ptr = dhrus,latits,PetTypeStrNames = PetsNams[1:nHrus],IdsHrus)
}
