% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dHRUMrun}
\alias{dHRUMrun}
\title{Run dHRUM and provides with model outputs}
\usage{
dHRUMrun(dHRUM_ptr)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}
}
\value{
list with matrix of caldata \code{[,1:4]} and hdata ts variables \code{[,5:27]} and names of vars
}
\description{
return list matrix with state variables and fluxes averaged over basin area.
}
\examples{
nHrus <- 2
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTInputsToAlldHrus(dhrus, Prec = prec, Temp = temp, as.Date("1990/01/30"))
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
                      SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
setParamsToAlldHrus(dHRUM_ptr = dhrus,as.numeric(ParDF[1,]),names(ParDF))
calcPetToAllHrus(dHRUM_ptr = dhrus,50.1,"HAMON")
outDta <- dHRUMrun(dHRUM_ptr = dhrus)
outDF <- data.frame(outDta$outDta)
names(outDF) <-c(outDta$VarsNams)
}
