% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gatherHBdata}
\alias{gatherHBdata}
\title{Calculates catchment spatially averaged hydrological balance}
\usage{
gatherHBdata(dHRUM_ptr)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRU instance}
}
\description{
Calculates catchment spatial average for fluxes and state variables
after the HB on every single HRU is updated.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
filname2 = "Calibrations/Amalie/indata/BP_1960_01_01.txt"
setInputsToAlldHrus(dHRUM_ptr = dhrus, filname2)
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
setParamsToAlldHrus(dHRUM_ptr = dhrus,as.numeric(ParDF[1,]),names(ParDF),TRUE,0)
calcPetToHrus(dHRUM_ptr = dhrus,50.1,"HAMON")
calcHBInAlldHrus(dHRUM_ptr = dhrus)
gatherHBdata(dHRUM_ptr = dhrus)
}
