% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.r
\name{get_Lai_Data}
\alias{get_Lai_Data}
\title{Read Lai Data from Package for Ohre river basin}
\usage{
get_Lai_Data()
}
\value{
A data frame containing the contents of 'LAI_POH_Category.rds' stored in .
}
\description{
This function locates and reads the sample LAI file file stored within the
package's internal data directory.
}
\examples{
\dontrun{
laiDT <- get_Lai_Data()
head(laiDT)
}
}
