% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCurdHRUpars}
\alias{getCurdHRUpars}
\title{Getting the current singeHMunit parameters.}
\usage{
getCurdHRUpars(dHRUM_ptr, singleHruId)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{singleHruId}{a Id of particular Hru}
}
\description{
shows the list of parameters for selected HRU
}
\examples{
nHrus <- 1
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTDateInputsToAlldHrus(dhrus, Prec = prec, Temp = temp,
  DateVec = as.Date(c("1990/01/30","1990/01/31","1990/02/01")))
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
 SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
setParamsToAlldHrus(dHRUM_ptr = dhrus,ParsVec = as.numeric(ParDF[1,]),ParsNames =names(ParDF))
getCurdHRUpars(dHRUM_ptr = dhrus,0)
}
