% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setFastResponsesToAlldHrus}
\alias{setFastResponsesToAlldHrus}
\title{Sets the types of surface retention models types to dHRU model for all single HRUs.}
\usage{
setFastResponsesToAlldHrus(dHRUM_ptr, fastResponseTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{fastResponseTypes}{a character vector of fast responses}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the fast response type to dHRUM to all HRUs. Possibe types: \code{SerialCascadeLinRes}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setFastResponsesToAlldHrus(dHRUM_ptr = dhrus,fastResponseTypes=rep("SerialCascadeLinRes",times= length(Areas)),hruIds=IdsHrus)
}
