% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setGWtypeToAlldHrus}
\alias{setGWtypeToAlldHrus}
\title{Sets the similar values of params to dHRU model for all single HRUs.}
\usage{
setGWtypeToAlldHrus(dHRUM_ptr, gwTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{gwTypes}{a charater vector of GW type names}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the groudnwater type of dHRUM equal to all HRUs. Possibe types: \code{LIN_RES} \code{LINL_RES} \code{LINBY_RES}
\code{POW_RES}, \code{EXP_RES} \code{LIN_2SE} \code{LIN_2PA} \code{FLEX_RES}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
setGWtypeToAlldHrus(dHRUM_ptr = dhrus)
}
