% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setInputsToAlldHrus}
\alias{setInputsToAlldHrus}
\title{Sets similar input data obtained from file to all single Hrus at dHRUM instance.}
\usage{
setInputsToAlldHrus(dHRUM_ptr, namInpFilet)
}
\arguments{
\item{namInpFilet}{a chacter vector to a single file with input data to dHRUM.}

\item{dHRU_ptr}{pointer to dHRU instance}
}
\description{
Loads the data from file to a single dHRU instance created  the \code{initdHruModel(nHrus,Areas,IdsHrus)} function
All iputs are same for each single HRU unit. File has on its first row YYYY MM DD,
remaining columns should have Temperature and Precipitation data.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
filname2 = "../../PDM/Development/PDM_dist/data/tests/inALL/BP_1960_01_01.txt"
setInputsToAlldHrus(dHRUM_ptr = dhrus, filname2)
}
