% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setInterceptiontypeToAlldHrus}
\alias{setInterceptiontypeToAlldHrus}
\title{Sets the types of interception models types to dHRU model for all single HRUs.}
\usage{
setInterceptiontypeToAlldHrus(dHRUM_ptr, intcptnTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{intcptnTypes}{a charater vector of Interception type names}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the interception type to dHRUM to all HRUs. Possibe types: \code{Rutter_Gash}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setInterceptiontypeToAlldHrus(dHRUM_ptr = dhrus,intcptnTypes=rep("Rutter_Gash",times= length(Areas)),hruIds=IdsHrus)
}
