% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setNumFastResAlldHrus}
\alias{setNumFastResAlldHrus}
\title{Sets the number of fast runoff reservoirs to dHRU model for all single HRUs.}
\usage{
setNumFastResAlldHrus(dHRUM_ptr, numFastRes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{numFastRes}{vector with integer numbers}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the number of reservoir for direct runoff on dHRUM  HRUs.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
ups = 1:nHrus
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setNumFastResAlldHrus(dHRUM_ptr = dhrus,numFastRes=ups,hruIds=IdsHrus)
}
