% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPondToOnedHru}
\alias{setPondToOnedHru}
\title{Sets the types of pond to dHRU model for one single HRU.}
\usage{
setPondToOnedHru(
  dHRUM_ptr,
  singleHruId,
  ValNames,
  ValVals,
  TypeNames,
  TypeVals
)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRU instance}

\item{singleHruId}{the numerical single HRU ID}

\item{ParsVec}{vector of values of parameters}

\item{ParsNames}{a charater vector of parameter names}
}
\description{
Setting of pond to one particular single HRU for dHRU.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTDateInputsToAlldHrus(dhrus, Prec = prec, Temp = temp,
DateVec = as.Date(c("1990/01/30","1990/01/31","1990/02/01")))
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
pondDF1 = data.frame( pondArea = 40500, PonsMax= 45000, MRF= 0.039)
pondDF2 = data.frame( ET = "ETpond1", in_SOISperc= "noPondSOISPerc", in_GWperc= "noPondGWPerc", out_SOISperc= "noPondSOISPerc", out_GWperc= "noPondGWPerc",regular_out="PondRouT3" )
HruPondID = 1
setPondToOnedHru(dHRUM_ptr = dhrus,HruPondID,names(pondDF1),as.numeric(pondDF1),as.character(pondDF2),names(pondDF2))
}
