% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPTDateInputsToAlldHrus}
\alias{setPTDateInputsToAlldHrus}
\title{Sets the similar Precipitation, Temperature, and Date vectors to dHRU.}
\usage{
setPTDateInputsToAlldHrus(dHRUM_ptr, Prec, Temp, DateVec)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRU instance}

\item{Prec}{vector of values of precipitation}

\item{Temp}{vector of values of temperature}

\item{DateVec}{vector of dates \code{as.Date("YYYY-MM-DD")}}
}
\description{
Setting the similar vectors of Precipitation and Temperature
to all single HRUs of dHRU. Loading the Date vector  from vector like variable.
}
\examples{
nHrus <- 2
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTDateInputsToAlldHrus(dhrus, Prec = prec, Temp = temp,
  DateVec = as.Date(c("1990/01/30","1990/01/31","1990/02/01")))
}
