% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPTInputsToAlldHrus}
\alias{setPTInputsToAlldHrus}
\title{Sets the similar Precipitation, Temperature vectors to dHRUM and init's the date using beg. of period.}
\usage{
setPTInputsToAlldHrus(dHRUM_ptr, Prec, Temp, inDate)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{Prec}{vector of values of precipitation}

\item{Temp}{vector of values of temperature}

\item{inDate}{the first date of simulation period}
}
\description{
Setting the similar vector of Precipitation and temperature to all single HRU.
Setting the calender using the first date fo period using the first date of period
}
\examples{
nHrus <- 2
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
setPTinputsToAlldHrus(dhrus, Prec = prec, Temp = temp, as.Date("1990/01/30"))
}
