% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPTInputsToAlldHrusFromFile}
\alias{setPTInputsToAlldHrusFromFile}
\title{Sets similar input data obtained from file to all single Hrus at dHRUM instance.}
\usage{
setPTInputsToAlldHrusFromFile(dHRUM_ptr, namInpFilet)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRU instance}

\item{namInpFilet}{a character vector to a single file with input data to dHRUM.}
}
\description{
Loads the data from file to a single dHRUM instance created  the \code{initdHruModel(nHrus,Areas,IdsHrus)} function
All inputs are same for each single HRU unit. File has on its first row YYYY MM DD,
remaining columns should have Temperature and Precipitation data.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
filname2 = "../Calibrations/Amalie/indata/BP_1960_01_01.txt"
setPTInputsToAlldHrusFromFile(dHRUM_ptr = dhrus, filname2)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
}
