% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setPTLInputsToAlldHrus}
\alias{setPTLInputsToAlldHrus}
\title{Sets the similar Precipitation, Temperature and LAI vectors to dHRUM and init's the date using beg. of period.}
\usage{
setPTLInputsToAlldHrus(dHRUM_ptr, Prec, Temp, Lai, inDate)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{Prec}{vector of values of precipitation}

\item{Temp}{vector of values of temperature}

\item{Lai}{vector of values of leaf area index}

\item{inDate}{the first date of simulation period}
}
\description{
Setting the similar vector of Precipitation ,temperature and LAI to all single HRU.
Setting the calender using the first date fo period using the first date of period
}
\examples{
nHrus <- 2
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
prec=c(1,2,3)
temp=c(1,2,3)
lai=c(1,2,3)
setPTLInputsToAlldHrus(dhrus, Prec = prec, Temp = temp, Lai = lai, as.Date("1990/01/30"))
}
