% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSnowMeltModeltypeToAlldHrus}
\alias{setSnowMeltModeltypeToAlldHrus}
\title{Sets the types of snow melt models types to dHRU model for all single HRUs.}
\usage{
setSnowMeltModeltypeToAlldHrus(dHRUM_ptr, snowMeltModelTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{snowMeltModelTypes}{a charater vector of Surface retention type names}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the snow melt type to dHRUM to all HRUs. Possibe types: \code{DFF}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setSnowMeltModeltypeToAlldHrus(dHRUM_ptr = dhrus,snowMeltModelTypes=rep("DDF",times= length(Areas)),hruIds=IdsHrus)
}
