% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSoilStorTypeToAlldHrus}
\alias{setSoilStorTypeToAlldHrus}
\title{Sets the similar values of params to dHRU model for all single HRUs.}
\usage{
setSoilStorTypeToAlldHrus(dHRUM_ptr, soilTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{soilTypes}{a charater vector of GW type names}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the Soil type of dHRUM equal to all HRUs. Possibe types: \code{PDM}; \code{PDM2}; \code{COLLIE_V2};
\code{NEW_ZEALAND}; \code{GR4J} ;\code{SBROOK_V1}; \code{HILLSLOPE}; \code{PLATEAU}
}
\details{
the \code{PDM} is the Moore implementation of soil PDM bucket with c_min and cmax

the \code{PDM2} is the Monatnari implementation of soil PDM bucket only with cmax (requires the setting of cmin to 0]
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setGWtypeToAlldHrus(dHRUM_ptr = dhrus,gwTypes=rep("LIN_2SE",times= length(Areas)),hruIds=IdsHrus)
setSoilStorTypeToAlldHrus(dHRUM_ptr = dhrus,soilTypes=rep("PDM",times= length(Areas)),hruIds=IdsHrus)
}
