% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSurfaceStortypeToAlldHrus}
\alias{setSurfaceStortypeToAlldHrus}
\title{Sets the types of surface retention models types to dHRU model for all single HRUs.}
\usage{
setSurfaceStortypeToAlldHrus(dHRUM_ptr, surfaceStorTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{surfaceStorTypes}{a charater vector of Surface retention type names}

\item{hruIds}{ids on Hrus}
}
\description{
Setting the surface retention type to dHRUM to all HRUs. Possibe types: \code{SurfaceAll}, \code{SurfacePRTL}
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
setSurfaceStortypeToAlldHrus(dHRUM_ptr = dhrus,surfaceStorTypes=rep("SurfaceAll",times= length(Areas)),hruIds=IdsHrus)
}
