% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initdHruModel}
\alias{initdHruModel}
\title{Initialization of dHRU pointer to a dHRUM}
\usage{
initdHruModel(dimdHru, vecAreas, hrusIDs)
}
\arguments{
\item{dimdHru}{a single \code{numberDta} number of single Hrus.}

\item{vecAreas}{a \code{numeric vector} of size \code{dimHru} of Areas for all single HRUs on dHRU.}

\item{hrusIDs}{a \code{character vector} of size \code{dimHru} of Id's for all single HRUs on dHRU.}
}
\value{
dHRU_ptr pointer to dHru instance.
}
\description{
Creates pointer instance of dHRUM for the catchment.
initializes a dimension of dHRUM controled by the number of single Hru, areas of all single Hrus,
and ID's of all single Hrus.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
}
