% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setGWtypeToAlldHrus}
\alias{setGWtypeToAlldHrus}
\title{Sets the similar values of params to dHRU model for all single HRUs.}
\usage{
setGWtypeToAlldHrus(dHRUM_ptr, gwTypes, hruIds)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRUM instance}

\item{gwTypes}{a charater vector of GW type names}

\item{dimdHru}{a single \code{numberDta} number of single Hrus.}
}
\description{
Setting of params to dHRUM.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
setGWtypeToAlldHrus(dHRUM_ptr = dhrus)
}
