% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setParsToDistdHRUM}
\alias{setParsToDistdHRUM}
\title{Sets the values of params to dHRU model with many Hrus.}
\usage{
setParsToDistdHRUM(dHRUM_ptr, ParsDF)
}
\arguments{
\item{dHRUM_ptr}{pointer to dHRU instance}

\item{ParsDF}{data.frame of parametrs cols show parameters, rows show the Hrus}
}
\description{
Setting of vectors of params to all HRUs for distributed dHRUM.
}
\examples{
nHrus <- 200
Areas <- runif(nHrus,min = 1,max  = 10)
IdsHrus <- paste0("ID",seq(1:length(Areas)))
dhrus <- initdHruModel(nHrus,Areas,IdsHrus)
filname2 = "../../PDM/Development/PDM_dist/data/tests/inALL/BP_1960_01_01.txt"
setInputsToAlldHrus(dHRUM_ptr = dhrus, filname2)
ParDF = data.frame( B_SOIL = 1.6, C_MAX = 100, B_EVAP = 2,  KS = 0.1, KF = 0.2, ADIV = 0.3, CDIV = 0.03,
SDIV = 0.03, CAN_ST = 2, STEM_ST = 2, CSDIV = 0.3, TETR = 5, DDFA = 0.5, TMEL = 0, RETCAP = 10 )
setParamsToOnedHru(dHRUM_ptr = dhrus,as.numeric(ParDF[1,]),names(ParDF),0)
}
